+++ sycl/doc/extensions/experimental/sycl_ext_oneapi_kernel_properties.asciidoc
@@ -115,6 +115,40 @@ supports.
  feature-test macro always has this value.
 |===
 
+=== Headers
+
+In the {dpcpp} implementation, the full API described in this extension is
+provided by:
+
+* `#include <sycl/ext/oneapi/kernel_properties/properties.hpp>`
+
+This umbrella header should be used when code needs property lists such as
+`properties{...}`, embedded kernel properties via `properties_tag`, or any of
+the non-launch kernel properties described in this extension.
+
+The implementation also provides lighter-weight headers for code that only
+needs free-function annotations:
+
+* `#include <sycl/ext/oneapi/kernel_properties/function_properties.hpp>` for
+  `nd_range_kernel<Dims>`, `single_task_kernel`, and
+  `SYCL_EXT_ONEAPI_FUNCTION_PROPERTY`.
+
+* `#include <sycl/ext/oneapi/kernel_properties/function_launch_properties.hpp>`
+  for the launch-configuration properties that can be applied directly to free
+  function declarations, including `work_group_size`,
+  `work_group_size_hint`, `sub_group_size`, `max_work_group_size`, and
+  `max_linear_work_group_size`.
+
+These lighter-weight headers are primarily intended for code paths where
+reduced compile time is a high priority and the application only needs
+function-level annotations.
+
+They retain validation of each individual property, but they do not provide
+the cross-property validation that depends on forming a full property list.
+For example, checks for conflicting combinations such as `work_group_size`
+together with `max_work_group_size` or `max_linear_work_group_size` are
+performed only when using the umbrella header.
+
 === Kernel Properties
 
 The kernel properties below correspond to kernel attributes defined in
