+++ sycl/test/extensions/properties/properties_kernel.cpp
@@ -60,6 +60,36 @@ int main() {
       is_property_value<decltype(max_work_group_size<8, 8, 8>)>::value);
   static_assert(
       is_property_value<decltype(max_linear_work_group_size<8>)>::value);
+  static_assert(
+      std::is_same_v<
+          std::remove_cv_t<decltype(nd_range_kernel<1>)>,
+          property_value<nd_range_kernel_key, std::integral_constant<int, 1>>>);
+  static_assert(std::is_same_v<std::remove_cv_t<decltype(single_task_kernel)>,
+                               property_value<single_task_kernel_key>>);
+  static_assert(
+      std::is_same_v<
+          std::remove_cv_t<decltype(work_group_size<2, 2>)>,
+          property_value<work_group_size_key, std::integral_constant<size_t, 2>,
+                         std::integral_constant<size_t, 2>>>);
+  static_assert(
+      std::is_same_v<std::remove_cv_t<decltype(work_group_size_hint<3, 3, 3>)>,
+                     property_value<work_group_size_hint_key,
+                                    std::integral_constant<size_t, 3>,
+                                    std::integral_constant<size_t, 3>,
+                                    std::integral_constant<size_t, 3>>>);
+  static_assert(
+      std::is_same_v<std::remove_cv_t<decltype(sub_group_size<7>)>,
+                     property_value<sub_group_size_key,
+                                    std::integral_constant<uint32_t, 7>>>);
+  static_assert(
+      std::is_same_v<std::remove_cv_t<decltype(max_work_group_size<8, 8>)>,
+                     property_value<max_work_group_size_key,
+                                    std::integral_constant<size_t, 8>,
+                                    std::integral_constant<size_t, 8>>>);
+  static_assert(
+      std::is_same_v<std::remove_cv_t<decltype(max_linear_work_group_size<8>)>,
+                     property_value<max_linear_work_group_size_key,
+                                    std::integral_constant<size_t, 8>>>);
 
   static_assert(
       std::is_same_v<work_group_size_key, decltype(work_group_size<8>)::key_t>);
@@ -107,9 +137,6 @@ int main() {
   static_assert(max_work_group_size<28, 29, 30>[1] == 29);
   static_assert(max_work_group_size<28, 29, 30>[2] == 30);
   static_assert(max_linear_work_group_size<28>.value == 28);
-
-  static_assert(std::is_same_v<decltype(sub_group_size<28>)::value_t,
-                               std::integral_constant<uint32_t, 28>>);
 
   singleAspectDeviceHasChecks<aspect::cpu>();
   singleAspectDeviceHasChecks<aspect::gpu>();
